+++ indexer/services/comlink/src/lib/restrict-countries.ts
@@ -7,6 +7,7 @@ import express from 'express';
 
 import { BlockedCode } from '../types';
 import { create4xxResponse } from './helpers';
+import { getIpAddr, isIndexerIp } from './utils';
 
 /**
  * Return an error code for users that access the API from a restricted country
@@ -16,6 +17,13 @@ export function rejectRestrictedCountries(
   res: express.Response,
   next: express.NextFunction,
 ) {
+  const ipAddr: string | undefined = getIpAddr(req);
+
+  // Don't enforce geo-blocking for internal IPs as they don't go through a proxy
+  if (ipAddr !== undefined && isIndexerIp(ipAddr)) {
+    return next();
+  }
+
   if (isRestrictedCountryHeaders(req.headers as CountryHeaders)) {
     return create4xxResponse(
       res,
